+++ src/modules/shipment/shipment.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { ShipmentService } from './shipment.service';
 import { ShipmentController } from './shipment.controller';
 import { MongooseModule } from '@nestjs/mongoose' ;
-import { RateSchema } from './entities/rate.entity';
-import { QuoteSchema } from './entities/quote.entity';
-import { ShipmentSchema } from './entities/shipment.entity';
+import { RateSchema } from './entities';
+import { QuoteSchema } from './entities';
+import { ShipmentSchema } from './entities';
 import { QuoteRepository, ShipmentRepository } from './repositories';
 import { RateRepository } from './repositories/rate.repository';
 import { BullModule } from '@nestjs/bull';
